+++ extensions/color-picker/src/pick-color.ts
@@ -1,13 +1,14 @@
 import { Clipboard, closeMainWindow, launchCommand, LaunchType, showHUD } from "@raycast/api";
 import { addToHistory } from "./history";
-import { PickColorCommandLaunchProps } from "./types";
-import { getFormattedColor, pickColor } from "./utils";
+import { PickColorCommandLaunchProps, Color } from "./types";
+import { getFormattedColor } from "./utils";
+import { pickColor } from "swift:../swift/color-picker";
 
 export default async function command(props: PickColorCommandLaunchProps) {
   await closeMainWindow();
 
   try {
-    const pickedColor = await pickColor();
+    const pickedColor = await pickColor<Color | undefined>();
     if (!pickedColor) {
       return;
     }