+++ extensions/search-npm/src/PackagListItem.tsx
@@ -34,11 +34,11 @@ interface PackageListItemProps {
 export interface Preferences {
   defaultOpenAction: 'openRepository' | 'openHomepage' | 'npmPackagePage'
   historyCount: string
+  showLinkToSearchResultsInListView: boolean
 }
 
 export const PackageListItem = ({
   result,
-  searchTerm,
   setHistory,
   isFavorited,
   handleFaveChange,
@@ -120,13 +120,15 @@ export const PackageListItem = ({
   }
 
   const accessories: List.Item.Accessory[] = [
-    {
-      icon: Icon.Tag,
-      tooltip: pkg?.keywords?.length ? pkg.keywords.join(', ') : '',
-    },
+    pkg?.keywords?.length
+      ? {
+          icon: Icon.Tag,
+          tooltip: pkg.keywords.join(', '),
+        }
+      : {},
   ]
   if (!isViewingFavorites) {
-    accessories.unshift(
+    accessories.push(
       {
         text: `v${pkg.version}`,
         tooltip: `Latest version`,
@@ -137,7 +139,7 @@ export const PackageListItem = ({
       },
     )
     if (isFavorited) {
-      accessories.unshift({
+      accessories.push({
         icon: Icon.Star,
       })
     }
@@ -170,12 +172,6 @@ export const PackageListItem = ({
                 return action
               })
               .filter(Boolean)}
-            {searchTerm ? (
-              <Action.OpenInBrowser
-                url={`https://www.npmjs.com/search?q=${searchTerm}`}
-                title="npm Search Results"
-              />
-            ) : null}
           </ActionPanel.Section>
           <ActionPanel.Section title="Actions">
             {isFavorited ? (