+++ extensions/vat-calculator/src/noview.jsx
@@ -1,5 +1,4 @@
-import { Clipboard, showHUD, getPreferenceValues } from "@raycast/api";
-import { runAppleScript } from "@raycast/utils";
+import { Clipboard, showHUD, getPreferenceValues, closeMainWindow } from "@raycast/api";
 
 import { numberWithVAT } from "./lib/vat";
 
@@ -7,12 +6,9 @@ export default async function quickVat(props) {
   var number = numberWithVAT(props.arguments.number);
 
   if (getPreferenceValues().autopaste) {
+    closeMainWindow();
+    Clipboard.paste(number);
     showHUD(`Pasted Result to Active App`);
-    await runAppleScript(`
-        tell application "System Events"
-            keystroke "${number}" 
-        end tell
-        `);
   } else {
     Clipboard.copy(number);
     showHUD("Copied Answer to Clipboard");