+++ extensions/ticktick/CHANGELOG.md
@@ -1,5 +1,7 @@
 # TickTick Changelog
 
+## [Fixed spelling error] - 2023-11-27
+
 ## [Added] - 2023-11-14
 
 - Add `Default Title` preference that allows you to set a default task title from your clipboard or selection when using the `Add Task` command.